// web/src/api/generated/ScheduleInput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MaintenanceTask } from "./MaintenanceTask";

export type ScheduleInput = { task: MaintenanceTask, cron: string, enabled: boolean, catalog_id: number, namespace: Array<string>, table: string, };