// web/src/api/generated/NamespaceDetail.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TableSummary } from "./TableSummary";

export type NamespaceDetail = { namespace: Array<string>, properties: { [key in string]?: string }, child_namespaces: Array<Array<string>>, tables: Array<TableSummary>, 
/**
 * True when only the first tables were summarized.
 */
truncated: boolean, };